% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{get_game_report}
\alias{get_game_report}
\title{Get Game Report}
\usage{
get_game_report(season, game_id)
}
\arguments{
\item{season}{Season in 8 numeric format (i.e. 20212022)}

\item{game_id}{Game Id in 5 digit format 20001. Regular season games begin with a 2, playoff games with a 3}
}
\value{
a list of tibbles with different scenarios (all, 5v5, 5v4, 4v5, etc)
}
\description{
Get a game report from Natural Stat Trick. Game reports are lists of data frames for different strength
scenarios, each showing things like coris for/against, fenwick for/against, xg for/against, goals for/against and toi
for the specific strenght.
}
