% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_season_to_file}
\alias{get_season_to_file}
\title{Get a Season and Write to File}
\usage{
get_season_to_file(
  season,
  playoffs = FALSE,
  filename = "~/Documents/natstattrick.csv"
)
}
\arguments{
\item{season}{season (four number year)}

\item{playoffs}{binary, whether or not to get the playoff data. Default false (get regular season). If true, only gets playoffs.}

\item{filename}{File to write data to (will also verify that data isn't in file before trying to download)}
}
\value{
invisible TRUE if finished ok.
}
\description{
a convenience function to write a whole season's nst data to file
}
