% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{nst_report}
\alias{nst_report}
\title{NST Short Report}
\usage{
nst_report(season, game_id)
}
\arguments{
\item{season}{Season in 8 numeric format (i.e. 20212022)}

\item{game_id}{Game Id in 5 digit format 20001. Regular season games begin with a 2, playoff games with a 3}
}
\value{
a tibble of all scenarios' data with
corsi, fenwick, shots, and more data returned for each team
}
\description{
Get a short report from NST for a single game.
For more information call get_game_report
}
