% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linescore.R
\name{nst_season_results}
\alias{nst_season_results}
\title{Linescore}
\usage{
nst_season_results(season, playoffs = FALSE)
}
\arguments{
\item{season}{Season in 8 numeric format (i.e. 20212022)}

\item{playoffs}{whether to get the playoff game data (TRUE) or regular season (FALSE)}
}
\description{
Get a whole season or playoffs at once!
}
