% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{nst_table_cleanup}
\alias{nst_table_cleanup}
\title{NST Table Cleanup}
\usage{
nst_table_cleanup(data)
}
\arguments{
\item{data}{partial nst table}
}
\value{
nst table after cleanup
}
\description{
NST Table Cleanup
}
