% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_game_df_to_file}
\alias{write_game_df_to_file}
\title{Write NST Game DF to R}
\usage{
write_game_df_to_file(
  gid,
  season = NULL,
  game_id = NULL,
  filename = "~/Documents/natstattrick.csv"
)
}
\arguments{
\item{gid}{gameID in YYYY0#GGGG format, or NULL to use season and game_id directly}

\item{season}{season in YYYYYYYY format (i.e. 20072008). Overwritten by calculated season if gid is provided}

\item{game_id}{game_id in ##### format (i.e. 21200). Overwritten by calculated game_id if gid is provided}

\item{file}{csv file to add the selected game to}
}
\value{
None, called for side-effects
}
\description{
Write NST Game DF to R
}
